requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
Pillow==12.1.1
python-telegram-bot==22.6
httpx==0.28.1
anyio==4.12.1

# Production
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
dj-database-url==2.2.0
